requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==2.0.0  --install-option="--without-c-extensions"  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django==3.1.10  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
# Extra line
# ------------------------------------------------------------------------------
google-api-python-client==1.4.1
django-admin-honeypot==1.1.0
django-bootstrap4==1.1.1
django-autocomplete-light==3.5.0
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==6.0.0
readtime==1.1.1
django-cookie-consent==0.2.6



# API NECESSARY APPS
# ------------------------------------------------------------------------------
djangorestframework-simplejwt==4.6.0 # Add proper token creation url links for app
drf-yasg==1.20.0 # Add documentation for api



# PHONE NUMBER VERIFICATION
# ------------------------------------------------------------------------------
django-phonenumber-field==5.0.0
phonenumbers==8.11.1


django-countries-plus==1.3.1 # https://pypi.org/project/django-countries-plus/
django-languages-plus==1.1.1 # https://pypi.org/project/django-languages-plus/

# SEARCH INDEX IMPORTS & PLUGINS
# ------------------------------------------------------------------------------
django-haystack==3.0
whoosh==2.7.4

# Admin
# ------------------------------------------------------------------------------
django-jet2==1.0.12
FeedParser==5.2.1
python-magic==0.4.15
django-ipware==3.0.1


# Minify HTML in django
# ------------------------------------------------------------------------------
# Todo: find the uses of these to minify wesbites and others to improve security
django-htmlmin==0.11.0
python-memcached==1.59


# Documentation API
# ------------------------------------------------------------------------------
coreapi==2.3.3 # https://github.com/core-api/python-client


# Add ajax commenting to django models
# ------------------------------------------------------------------------------
# Todo: find the uses of these to minify wesbites and others to improve security
django-comments-dab==2.6.0

# Add swagger for friendlier api view
# ------------------------------------------------------------------------------
django-rest-swagger==2.2.0

# Add image resize for models
# ------------------------------------------------------------------------------
django-resized==0.3.11
flower==0.9.7


# Add geolocation to django app
# ------------------------------------------------------------------------------
geoip2==4.1.0
maxminddb==2.0.3


# use x-editable to work with django 
# inline edit properties of a detail view
# https://stackoverflow.com/questions/31852816/save-changes-to-database-using-django-and-x-editable
# provies solution to implement this easily